fix(v1.8): extrai tipos FIRE para src/types/fire.ts — corrige runtime error
Componentes client importavam tipos de API routes que importam
server-only (prisma, auth). Next.js resolvia o módulo como undefined
causando "Element type is invalid" na página /independencia.

Solução: todos os tipos compartilhados (FireResponse, WealthItemSerialized,
etc.) movidos para src/types/fire.ts sem dependências server-only.
Route files e client components agora importam de @/types/fire.
